requirements.txt
Scrapy~=2.6.2
pandas==1.5.0
tksheet~=5.4.1
PyYAML~=6.0